worked on comments by ash
